% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{dist_hist}
\alias{dist_hist}
\title{Plot Histogram of a Distribution}
\usage{
dist_hist(dist_f, ns = 1000, theme = "theme_minimal", bins = 30, seed = 42)
}
\arguments{
\item{dist_f}{a quoted distribution function: rnorm, runif or rcauchy}

\item{ns}{how many numbers should be generated?}

\item{theme}{a quoted ggplot theme to be used}

\item{bins}{how many bins should be used in the histogram?}

\item{seed}{passed to \code{set.seed}}
}
\value{
histogram of a sample of samplesize \code{ns} from a distribution
    \code{dist_f} using ggplot \code{theme}
}
\description{
This function uses non standard evaluation (NSE) to create a
    plot title.
}
\examples{
dist_hist("runif", ns = 100, theme = "theme_classic")
# Code
dist_hist
}
