% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dropdown_to_filter.R
\name{dropdown_to_filter}
\alias{dropdown_to_filter}
\title{Dropdown Filter Example}
\usage{
dropdown_to_filter(flower)
}
\arguments{
\item{flower}{value in \code{unique(iris$Species)}}
}
\value{
ggplot object
}
\description{
Dropdown Filter Example
}
\examples{
dropdown_to_filter("setosa")
# Code
dropdown_to_filter
}
