% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/number_to_sample.R
\name{number_to_sample}
\alias{number_to_sample}
\title{Number to Sample Size Examle}
\usage{
number_to_sample(n)
}
\arguments{
\item{n}{a number}
}
\value{
a plot
}
\description{
Number to Sample Size Examle
}
\examples{
number_to_sample(42)
# Code
number_to_sample
}
