% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_to_title.R
\name{text_to_title}
\alias{text_to_title}
\title{Text to Title Example}
\usage{
text_to_title(plot_title)
}
\arguments{
\item{plot_title}{Text.}
}
\value{
ggplot object
}
\description{
Text to Title Example
}
\examples{
text_to_title("My Title")
# Code
text_to_title
}
