% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sepals_and_petals.R
\name{sepals_and_petals}
\alias{sepals_and_petals}
\title{Plot Sepals and Petals in Iris}
\usage{
sepals_and_petals(x, y)
}
\arguments{
\item{x}{value in \code{names(iris)[1:4]}}

\item{y}{value in \code{names(iris)[1:4]}}
}
\value{
ggplot object
}
\description{
This function can plot any of the numeric columns in the iris
    dataset against each other.
}
\examples{
sepals_and_petals("Sepal.Length", "Petal.Length")
sepals_and_petals("Petal.Length", "Sepal.Length")

}
